refactor: inline overkill custom hook
