+++ packages/twenty-front/src/modules/workflow/components/RightDrawerWorkflowSelectAction.tsx
@@ -1,9 +1,12 @@
 import { RightDrawerWorkflowSelectActionContent } from '@/workflow/components/RightDrawerWorkflowSelectActionContent';
-import { useFindShowPageWorkflow } from '@/workflow/hooks/useFindShowPageWorkflow';
+import { useWorkflowWithCurrentVersion } from '@/workflow/hooks/useWorkflowWithCurrentVersion';
+import { showPageWorkflowIdState } from '@/workflow/states/showPageWorkflowIdState';
+import { useRecoilValue } from 'recoil';
 import { isDefined } from 'twenty-ui';
 
 export const RightDrawerWorkflowSelectAction = () => {
-  const workflow = useFindShowPageWorkflow();
+  const showPageWorkflowId = useRecoilValue(showPageWorkflowIdState);
+  const workflow = useWorkflowWithCurrentVersion(showPageWorkflowId);
 
   if (!isDefined(workflow)) {
     return null;
